requirements.txt
elevenlabs>=0.2.26
openai>=1.3.0
python-dotenv>=1.0.0
SpeechRecognition>=3.10.0
pyaudio>=0.2.13
sounddevice>=0.4.6
numpy>=1.24.0
spotipy>=2.23.0